Added #INSTRUM param to song files
Added an optional #INSTRUM parameter to the song files. It allows to specify an additional audio file for the instrumental variation of the song, to be played when the F7 key is pressed instead of the #MP3 file with the voice filter applied. This allows for a much higher quality no-voice karaoke on the songs where this instrumental variation is easily found (like with a lot of modern film OSTs). If the parameter is not found in the song file, the F7 key just activates the filter as it did before (no change with previous behaviour). Only if the parameter is in the song file and the file it points is a valid audio file, it will play that file instead of enabling the filter over the #MP3 file. Adding the parameter to a song file doesn't break old versions of the game, because the parameter is just ignored.